refactor: Chat.jsx 컴포넌트 분리
